+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Both client libraries are pre-1.0, and they have separate versioning.
 - Added support for more parameters to `Sandbox.create()`:
   - `blockNetwork`: Whether to block all network access from the sandbox.
   - `cidrAllowlist`: List of CIDRs the sandbox is allowed to access.
+  - `gpu`: GPU reservation for the sandbox (e.g. "A100", "T4:2", "A100-80GB:4").
 
 ## modal-js/v0.3.15, modal-go/v0.0.15